Database/requirements.txt
mysql-connector-python==8.0.26
protobuf==3.18.0
python-dotenv==0.19.0
